% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainGoel.R
\name{MainGoel}
\alias{MainGoel}
\title{Algoritmo MCMC Poisson Process (Goel)}
\usage{
MainGoel(data, sites, iter = 410000, bar = 4e+05)
}
\arguments{
\item{data}{Matriz de dados simulados de um processo de Poisson}

\item{sites}{Malha de pontos}

\item{iter}{número inteiro positivo, interações do algoritmo MCMC}

\item{bar}{número inteiro positivo, período de burn-in do algoritmo MCMC}
}
\description{
Esta é a função para realizar o algoritmo Monte Carlo via Cadeias de Markov para os parâmetros de um processo de Poisson.
}
